requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiokafka==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
fastapi==0.115.12
filelock==3.13.4
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-ollama==0.3.2
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.37
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
ollama==0.4.8
openai==1.76.0
orjson==3.10.16
packaging==24.2
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdfium2==4.30.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==69.5.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.25.3
watchfiles==1.0.5
websockets==14.2
yarl==1.20.0
zstandard==0.23.0
python-docx>=0.8.11
